requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
argcomplete==2.0.0
async-timeout==4.0.2
attrs==23.1.0
bitarray==2.7.6
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.1.0
cycler==0.11.0
cytoolz==0.12.1
eth-abi==4.1.0
eth-account==0.9.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.4.0
eth-utils==2.1.1
fonttools==4.40.0
frozenlist==1.3.3
hexbytes==0.3.1
idna==3.4
importlib-resources==5.12.0
jsonschema==4.17.3
kiwisolver==1.4.4
lru-dict==1.2.0
matplotlib==3.7.1
multidict==6.0.4
numpy==1.25.0
packaging==23.1
pandas==2.0.2
parsimonious==0.9.0
Pillow==9.5.0
pipx==1.0.0
protobuf==4.23.3
pycryptodome==3.18.0
pyparsing==3.1.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pyzmq==25.1.0
regex==2023.6.3
requests==2.31.0
rlp==3.0.0
six==1.16.0
toolz==0.12.0
tzdata==2023.3
urllib3==2.0.3
userpath==1.8.0
web3==6.5.0
websockets==11.0.3
yarl==1.9.2
zipp==3.15.0
